Extract InsertAtEvaluator to its own file
